requirements.txt
# 股癌 Digest Pipeline 依賴

# AI Providers（三選一自動 fallback）
anthropic>=0.40.0
openai>=1.30.0
google-genai>=0.1.0

feedparser>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Stock data
yfinance>=0.2.48
pandas>=2.0.0



